refactor(club-detail): ShareButton 제거 및 FloatingButtonGroup 클럽 공유 로직 통합
- 미사용 ShareButton 컴포넌트 및 share_icon SVG 파일 삭제
- FloatingButtonGroup에서 클럽 상세 페이지 감지 후 클럽 특화 공유 로직 적용
- 클럽 외 페이지는 기존 window.location.href fallback 유지
