--- frontend/src/pages/ClubDetailPage/components/ShareButton/ShareButton.styles.ts
@@ -1,18 +1,0 @@
-import styled from 'styled-components';
-import { media } from '@/styles/mediaQuery';
-
-export const ShareButtonContainer = styled.div`
-  display: flex;
-  justify-content: flex-end;
-  cursor: pointer;
-`;
-
-export const ShareButtonIcon = styled.img`
-  width: 60px;
-  height: 60px;
-
-  ${media.mobile} {
-    width: 44px;
-    height: 44px;
-  }
-`;
